# bin/Release/net5.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "bright_web_api",
    "version": "v1"
  },
  "paths": {
    "/api/Repairs/add-repair": {
      "post": {
        "tags": [
          "Repairs"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/RepairVM"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Repairs/update-repair-by-id/{id}": {
      "put": {
        "tags": [
          "Repairs"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/RepairVM"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Repairs/delete-repair-by-id/{id}": {
      "delete": {
        "tags": [
          "Repairs"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Repairs/get-all-repairs": {
      "get": {
        "tags": [
          "Repairs"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Repairs/get-repair-by-id/{id}": {
      "get": {
        "tags": [
          "Repairs"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Steps/add-step": {
      "post": {
        "tags": [
          "Steps"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/StepVM"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Steps/get-all-steps": {
      "get": {
        "tags": [
          "Steps"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Steps/get-step-by-id/{id}": {
      "get": {
        "tags": [
          "Steps"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Steps/get-steps-by-repair-id/{repairId}": {
      "get": {
        "tags": [
          "Steps"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "repairId",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Tools/add-tool": {
      "post": {
        "tags": [
          "Tools"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/ToolVM"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Tools/get-all-tools": {
      "get": {
        "tags": [
          "Tools"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Tools/get-tool-by-id/{id}": {
      "get": {
        "tags": [
          "Tools"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    }
  },
  "definitions": {
    "RepairVM": {
      "type": "object",
      "properties": {
        "title": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "image": {
          "type": "string"
        },
        "difficulty": {
          "format": "int32",
          "type": "integer"
        },
        "time": {
          "format": "int32",
          "type": "integer"
        },
        "pdf": {
          "type": "string"
        }
      }
    },
    "StepVM": {
      "type": "object",
      "properties": {
        "title": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "image": {
          "type": "string"
        },
        "video": {
          "type": "string"
        },
        "repairId": {
          "format": "int32",
          "type": "integer"
        }
      }
    },
    "ToolVM": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "image": {
          "type": "string"
        }
      }
    }
  }
}